뒤로가기 버튼 막기와 메인에서 뒤로가기 두 번 누를 시 앱 종료
뒤로가기 버튼 막기와 메인에서 뒤로가기 두 번 누를 시 앱 종료 일기 삭제 다이얼로드

+++ app/src/main/java/com/example/picturediary/TimelineActivity.kt
@@ -241,17 +241,27 @@ class TimelineActivity : AppCompatActivity() {
             }
 
             viewHolder.like_number.text = "좋아요 " + contentDTOs[p1].favoriteCount + "개"
+
             viewHolder.delete_picture.setOnClickListener {
                 deleteDiary(contentDTOs[p1].contentId)
             }
         }
 
+        //일기 삭제
         private fun deleteDiary(ContentID: String?) {
-            if (ContentID != null) {
-                firestore.collection("contents")
-                    .document(ContentID)
-                    .delete()
-            }
+            val builder = AlertDialog.Builder(this@TimelineActivity)
+            builder.setTitle("알림")
+                .setMessage("일기를 그룹에서 삭제하시겠습니까?")
+                .setPositiveButton("삭제", DialogInterface.OnClickListener{dialog, id ->
+                    if (ContentID != null) {
+                        firestore.collection("contents")
+                            .document(ContentID)
+                            .delete()
+                    }
+                })
+                .setNegativeButton("취소",DialogInterface.OnClickListener { dialog, id ->})
+            builder.show()
+
         }
 
         override fun getItemCount(): Int {
